Added another ship type
This is to supplement carrot, so ship type has five options that do
things. It’s a mode that turns the game into Ikaruga. You’re only hurt
by things that AREN”T your color and absorb shots of things that are
your color. When you absorb shots you power up a “bomb”.